refactor: create a shared TemplateTags component
